+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 # HidingBar - Classic
 
-* Updated options with slider.
-* Updated deDE translation.
+* MiniMapTrackingFrame is grabbed now.